fix : 댓글에 알림함수 추가, authorId => userId, 댓글 조회 시간 순으로 변경

See merge request sw_track/class_07/web_project_2/team8/server!27

+++ src/services/commentService.js
@@ -4,21 +4,23 @@ import Comment from '../models/schemas/Comment.js';
 import Reply from '../models/schemas/Reply.js';
 
 // 1. 마이페이지에서 내가 썼던 댓글 조회
-export async function getAllCommentsByUserId(authorId) {
-  const comments = await Comment.find({ authorId }).catch((error) => {
+export async function getAllCommentsByUserId(userId) {
+  const comments = await Comment.find({ authorId: userId }).catch((error) => {
     throw new CustomError(commonError.DB_ERROR, 'Internal server error', {
       statusCode: 500,
       cause: error,
     });
   });
-  const replies = await Reply.find({ authorId }).catch((error) => {
+  const replies = await Reply.find({ authorId: userId }).catch((error) => {
     throw new CustomError(commonError.DB_ERROR, 'Internal server error', {
       statusCode: 500,
       cause: error,
     });
   });
 
-  return [...comments, ...replies];
+  const myComments = [...comments, ...replies];
+  myComments.sort((a, b) => a.createdAt - b.createdAt);
+  return myComments;
 }
 
 // 2. 게시물에 있는 댓글 조회
@@ -36,7 +38,7 @@ export async function getCommentsByPostId(postId) {
 }
 
 // 3. 특정 게시물에 댓글 작성
-export async function createComment(authorId, postId, content, parentComment) {
+export async function createComment(userId, postId, content, parentComment) {
   // 부모 댓글이 있는 경우 => 대댓글 작성
   if (parentComment) {
     const parentCommentId = await Comment.findById(parentComment).catch((error) => {
@@ -51,7 +53,7 @@ export async function createComment(authorId, postId, content, parentComment) {
       });
     }
 
-    const newReply = new Reply({ parentComment: parentCommentId, authorId, postId, content });
+    const newReply = new Reply({ parentComment: parentCommentId, authorId: userId, postId, content });
     const reply = await newReply.save();
     parentCommentId.reply.push(reply._id);
     await parentCommentId.save();
@@ -59,7 +61,7 @@ export async function createComment(authorId, postId, content, parentComment) {
   }
 
   // 부모 댓글이 없는 경우 => 첫번째 댓글 생성
-  const newComment = new Comment({ authorId, postId, content });
+  const newComment = new Comment({ authorId: userId, postId, content });
   const comment = await newComment.save().catch((error) => {
     throw new CustomError(commonError.DB_ERROR, 'Internal server error', {
       statusCode: 500,
@@ -70,8 +72,8 @@ export async function createComment(authorId, postId, content, parentComment) {
   return comment;
 }
 
-// 4. 특정 게시물에 작성한 댓글 수정, (authorId를 userId로 변경)
-export async function updateComment(commentId, authorId, content) {
+// 4. 특정 게시물에 작성한 댓글 수정
+export async function updateComment(commentId, userId, content) {
   const foundComment = await Comment.findById(commentId);
   const foundReply = await Reply.findById(commentId);
 
@@ -85,7 +87,7 @@ export async function updateComment(commentId, authorId, content) {
   const comment = isReply ? Reply : Comment;
   const userIdTest = isReply ? foundReply : foundComment;
 
-  if (!userIdTest.authorId.equals(authorId)) {
+  if (!userIdTest.authorId.equals(userId)) {
     // comment.authorId = 댓글 작성자 ID, authorId = 사용자 ID
     throw new CustomError(commonError.USER_MATCH_ERROR, '댓글을 수정할 권한이 없습니다.', {
       statusCode: 403,
@@ -109,15 +111,26 @@ export async function updateComment(commentId, authorId, content) {
 }
 
 // 5. 특정 게시물에 작성한 댓글 삭제
-export async function deleteComment(authorId, commentId) {
+export async function deleteComment(userId, commentId) {
   const isReply = await Reply.findById(commentId);
 
+  let deletedComment;
   if (isReply) {
     // 대댓글 : Comment스키마의 reply필드에서, Reply스키마에서 자신의 _id 제거
-    await Comment.findByIdAndUpdate(isReply.parentComment, { $pull: { reply: commentId } }, { new: true });
-    await Reply.deleteOne({ _id: commentId, authorId });
+    deletedComment = await Comment.findByIdAndUpdate(
+      isReply.parentComment,
+      { $pull: { reply: commentId } },
+      { new: true },
+    );
+    await Reply.deleteOne({ _id: commentId, authorId: userId });
   } else {
-    await Comment.deleteOne({ _id: commentId, authorId });
+    deletedComment = await Comment.deleteOne({ _id: commentId, authorId: userId });
+  }
+
+  if (deletedComment.deletedCount === 0) {
+    throw new CustomError(commonError.COMMENT_DELETE_ERROR, '댓글 삭제를 실패하였습니다.', {
+      statusCode: 404,
+    });
   }
 
   return { message: '댓글이 삭제되었습니다.' };
